move dash inside notebook
